% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_and_as_hmsc.R
\name{as_hmsc_random}
\alias{as_hmsc_random}
\title{Build historical random-effect table}
\usage{
as_hmsc_random(hist, strategy = c("minimal", "balanced", "full"))
}
\arguments{
\item{hist}{An \code{hmsc_history_indices} object.}

\item{strategy}{Variable selection strategy.}
}
\value{
A site-level data.frame or NULL.
}
\description{
Returns selected categorical historical variables that may be used in HMSC
\code{studyDesign} and \code{ranLevels}.
}
