% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_standard_tables.R
\name{build_history_events}
\alias{build_history_events}
\title{Build a history event table from standardized events}
\usage{
build_history_events(raw_events)
}
\arguments{
\item{raw_events}{Standardized event table.}
}
\value{
A dispersal/loss event data.frame or NULL.
}
\description{
Build a history event table from standardized events
}
