% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_standard_tables.R
\name{build_speciation_events}
\alias{build_speciation_events}
\title{Build a speciation event table from standardized events}
\usage{
build_speciation_events(raw_events, phy = NULL)
}
\arguments{
\item{raw_events}{Standardized event table.}

\item{phy}{Optional tree. Not used in v0.1.}
}
\value{
A speciation event data.frame or NULL.
}
\description{
Build a speciation event table from standardized events
}
