% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dispersal_sources.R
\name{calc_dispersal_source_indices}
\alias{calc_dispersal_source_indices}
\title{Calculate historical dispersal source indices}
\usage{
calc_dispersal_source_indices(
  proj,
  probability_weighted = TRUE,
  shannon_base = exp(1)
)
}
\arguments{
\item{proj}{A \code{HmscEcoEvo_data} object.}

\item{probability_weighted}{Logical; use entry probabilities when available.}

\item{shannon_base}{Base for Shannon diversity.}
}
\value{
A list with \code{XData}, \code{random}, and \code{diagnostics}.
}
\description{
Calculate historical dispersal source indices
}
