% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_speciation.R
\name{calc_speciation_indices}
\alias{calc_speciation_indices}
\title{Calculate speciation-related historical indices}
\usage{
calc_speciation_indices(proj, probability_weighted = TRUE)
}
\arguments{
\item{proj}{A \code{HmscEcoEvo_data} object.}

\item{probability_weighted}{Logical; use event probabilities when available.}
}
\value{
A list with \code{XData} and \code{diagnostics}.
}
\description{
Calculate speciation-related historical indices
}
