% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trait_history_indices.R
\name{calc_trait_history_indices}
\alias{calc_trait_history_indices}
\title{Calculate species-level trait history indices}
\usage{
calc_trait_history_indices(proj, probability_weighted = TRUE)
}
\arguments{
\item{proj}{A \code{HmscEcoEvo_data} object.}

\item{probability_weighted}{Logical; use probabilities where available.}
}
\value{
A list with \code{TrData} and \code{diagnostics}.
}
\description{
Calculate species-level trait history indices
}
