% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ecoevo.R
\name{extract_hmsc_ecoevo}
\alias{extract_hmsc_ecoevo}
\title{Extract HMSC outputs into \code{hmsc_ecoevo}}
\usage{
extract_hmsc_ecoevo(model, ...)
}
\arguments{
\item{model}{A fitted Hmsc model.}

\item{...}{Additional arguments passed to \code{\link[=hmsc_ecoevo]{hmsc_ecoevo()}}.}
}
\value{
A \code{hmsc_ecoevo} object.
}
\description{
Extract HMSC outputs into \code{hmsc_ecoevo}
}
