% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_check_comm}
\alias{hee_check_comm}
\title{Check a community matrix}
\usage{
hee_check_comm(
  comm,
  response_type = c("presence", "count", "abundance", "cover")
)
}
\arguments{
\item{comm}{Site by species community matrix/data.frame.}

\item{response_type}{Response type.}
}
\value{
A diagnostic list.
}
\description{
Check a community matrix
}
