% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_check_timecube}
\alias{hee_check_timecube}
\title{Check that a time cube contains required variables and ages}
\usage{
hee_check_timecube(
  env_cube,
  variables = NULL,
  times = NULL,
  landmask = "land_mask_dem"
)
}
\arguments{
\item{env_cube}{A time cube.}

\item{variables}{Required variables.}

\item{times}{Required ages in Ma.}

\item{landmask}{Optional land mask variable to require.}
}
\value{
A diagnostic list.
}
\description{
Check that a time cube contains required variables and ages
}
