% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_check_traits}
\alias{hee_check_traits}
\title{Check a trait table}
\usage{
hee_check_traits(traits, species)
}
\arguments{
\item{traits}{Species by trait data.frame.}

\item{species}{Species names.}
}
\value{
A diagnostic list.
}
\description{
Check a trait table
}
