% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_earth_land_age}
\alias{hee_earth_land_age}
\title{Canonical land-age driver}
\usage{
hee_earth_land_age(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=hee_land_age]{hee_land_age()}}.}
}
\value{
Output from \code{\link[=hee_land_age]{hee_land_age()}}.
}
\description{
Preferred driver-layer name for \code{\link[=hee_land_age]{hee_land_age()}}.
}
