% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_earth_landscape_events}
\alias{hee_earth_landscape_events}
\title{Canonical landscape-event driver}
\usage{
hee_earth_landscape_events(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=hee_landscape_events]{hee_landscape_events()}}.}
}
\value{
Output from \code{\link[=hee_landscape_events]{hee_landscape_events()}}.
}
\description{
Preferred driver-layer name for \code{\link[=hee_landscape_events]{hee_landscape_events()}}.
}
