% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_environmental_filtering_suitability}
\alias{hee_environmental_filtering_suitability}
\title{Canonical environmental filtering suitability function}
\usage{
hee_environmental_filtering_suitability(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=hee_lineage_suitability]{hee_lineage_suitability()}}.}
}
\value{
Output from \code{\link[=hee_lineage_suitability]{hee_lineage_suitability()}}.
}
\description{
Six-process name for \code{\link[=hee_lineage_suitability]{hee_lineage_suitability()}}.
}
