% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_evolution_ancestral_response}
\alias{hee_evolution_ancestral_response}
\title{Canonical evolution ancestral response}
\usage{
hee_evolution_ancestral_response(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=hee_trait_mediated_ancestral_response]{hee_trait_mediated_ancestral_response()}}.}
}
\value{
Output from \code{\link[=hee_trait_mediated_ancestral_response]{hee_trait_mediated_ancestral_response()}}.
}
\description{
Six-process name for \code{\link[=hee_trait_mediated_ancestral_response]{hee_trait_mediated_ancestral_response()}}.
}
