% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geological_processes.R
\name{hee_gamma_shift}
\alias{hee_gamma_shift}
\title{Compute Gamma shift diagnostics}
\usage{
hee_gamma_shift(evo, ...)
}
\arguments{
\item{evo}{A \code{hmsc_ecoevo} object.}

\item{...}{Arguments passed to \code{\link[=calc_gamma_evolution_metrics]{calc_gamma_evolution_metrics()}}.}
}
\value{
A \code{hmsc_gamma_evolution_metrics} object.
}
\description{
Workflow alias for \code{\link[=calc_gamma_evolution_metrics]{calc_gamma_evolution_metrics()}}.
}
