% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_earth_biota_core.R
\name{hee_global_grid_dynamic_formula_catalog}
\alias{hee_global_grid_dynamic_formula_catalog}
\title{Formula catalog for no-BioGeoBEARS global grid dynamics}
\usage{
hee_global_grid_dynamic_formula_catalog()
}
\value{
Data frame describing the formulas used by
\code{\link[=hee_global_grid_dynamic_occupancy]{hee_global_grid_dynamic_occupancy()}}.
}
\description{
Formula catalog for no-BioGeoBEARS global grid dynamics
}
