% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_history_data}
\alias{hee_history_data}
\alias{hee_history_build_tables}
\alias{hee_history_standardize}
\alias{hee_history_indices}
\alias{hee_history_as_xdata}
\alias{hee_history_as_trdata}
\alias{hee_history_as_random}
\alias{hee_history_detect_leakage}
\title{Canonical historical-predictor data constructor}
\usage{
hee_history_data(...)

hee_history_build_tables(...)

hee_history_standardize(...)

hee_history_indices(...)

hee_history_as_xdata(...)

hee_history_as_trdata(...)

hee_history_as_random(...)

hee_history_detect_leakage(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=HmscEcoEvo_data]{HmscEcoEvo_data()}}.}
}
\value{
Output from \code{\link[=HmscEcoEvo_data]{HmscEcoEvo_data()}}.
}
\description{
Preferred history-layer name for \code{\link[=HmscEcoEvo_data]{HmscEcoEvo_data()}}.
}
