% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_make_paleo_grid}
\alias{hee_make_paleo_grid}
\title{Extract a paleoenvironmental grid for one or more ages}
\usage{
hee_make_paleo_grid(
  env_cube,
  times,
  variables,
  landmask = "land_mask_dem",
  land_only = TRUE
)
}
\arguments{
\item{env_cube}{Time cube.}

\item{times}{Ages in Ma.}

\item{variables}{Variables to extract.}

\item{landmask}{Land mask variable.}

\item{land_only}{Logical; keep only land cells with landmask > 0.}
}
\value{
A data.frame.
}
\description{
Extract a paleoenvironmental grid for one or more ages
}
