% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geological_processes.R
\name{hee_phylo_niche_signal}
\alias{hee_phylo_niche_signal}
\title{Compute phylogenetic niche signal}
\usage{
hee_phylo_niche_signal(evo, ...)
}
\arguments{
\item{evo}{A \code{hmsc_ecoevo} object.}

\item{...}{Arguments passed to \code{\link[=calc_phylo_signal_metrics]{calc_phylo_signal_metrics()}}.}
}
\value{
A \code{hmsc_phylo_signal_metrics} object.
}
\description{
Workflow alias for \code{\link[=calc_phylo_signal_metrics]{calc_phylo_signal_metrics()}}.
}
