% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_project_hmsc_table}
\alias{hee_project_hmsc_table}
\title{Project HMSC Beta responses for an arbitrary paleo predictor table}
\usage{
hee_project_hmsc_table(
  evo = NULL,
  beta = NULL,
  X_past,
  recipe = NULL,
  id_cols = c("site_id", "track_id", "time_ma", "matched_time_ma", "lon", "lat",
    "paleo_lon", "paleo_lat", "cell_id"),
  species = NULL,
  link = NULL,
  intercept_axis = "(Intercept)"
)
}
\arguments{
\item{evo}{Optional \code{hmsc_ecoevo} object.}

\item{beta}{Optional species by predictor matrix.}

\item{X_past}{Paleo predictor table.}

\item{recipe}{Optional projection recipe.}

\item{id_cols}{Columns to keep in the long output.}

\item{species}{Optional species subset.}

\item{link}{Link/response scale.}

\item{intercept_axis}{Intercept column in Beta.}
}
\value{
A long \code{hee_suitability} table.
}
\description{
Project HMSC Beta responses for an arbitrary paleo predictor table
}
