% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_refugia}
\alias{hee_refugia}
\title{Summarize long-term refugia potential}
\usage{
hee_refugia(richness, group_col = NULL)
}
\arguments{
\item{richness}{Output from \code{\link[=hee_richness]{hee_richness()}}.}

\item{group_col}{Spatial or track grouping column.}
}
\value{
A cell-level refugia summary.
}
\description{
Summarize long-term refugia potential
}
