% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{hee_render_word_report}
\alias{hee_render_word_report}
\title{Render a Word report from an R Markdown file}
\usage{
hee_render_word_report(rmd, output_file, params = list())
}
\arguments{
\item{rmd}{Path to an R Markdown report.}

\item{output_file}{Output DOCX path.}

\item{params}{Optional R Markdown parameters.}
}
\value{
Output DOCX path.
}
\description{
Render a Word report from an R Markdown file
}
