% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_naming.R
\name{hee_speciation_events}
\alias{hee_speciation_events}
\title{Canonical speciation event standardisation}
\usage{
hee_speciation_events(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=build_speciation_events]{build_speciation_events()}}.}
}
\value{
Output from \code{\link[=build_speciation_events]{build_speciation_events()}}.
}
\description{
Six-process name for \code{\link[=build_speciation_events]{build_speciation_events()}}.
}
