% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geological_processes.R
\name{hee_trait_mediation_signal}
\alias{hee_trait_mediation_signal}
\title{Compute trait-mediated niche signal}
\usage{
hee_trait_mediation_signal(evo, ...)
}
\arguments{
\item{evo}{A \code{hmsc_ecoevo} object.}

\item{...}{Arguments passed to \code{\link[=calc_trait_mediation_metrics]{calc_trait_mediation_metrics()}}.}
}
\value{
A \code{hmsc_trait_mediation_metrics} object.
}
\description{
Workflow alias for \code{\link[=calc_trait_mediation_metrics]{calc_trait_mediation_metrics()}}.
}
