% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_time_workflow.R
\name{hee_turnover}
\alias{hee_turnover}
\title{Calculate community turnover between adjacent time slices}
\usage{
hee_turnover(projection)
}
\arguments{
\item{projection}{Projection table.}
}
\value{
A data.frame with Bray-Curtis-like turnover.
}
\description{
Calculate community turnover between adjacent time slices
}
