% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_history_indices.R
\name{history_dictionary}
\alias{history_dictionary}
\title{History index dictionary}
\usage{
history_dictionary()
}
\value{
A data.frame describing main variables.
}
\description{
History index dictionary
}
