% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoevo_utils.R
\name{print.hmsc_ecoevo_plot_list}
\alias{print.hmsc_ecoevo_plot_list}
\title{Print a HmscEcoEvo plot list}
\usage{
\method{print}{hmsc_ecoevo_plot_list}(x, ...)
}
\arguments{
\item{x}{A plot-list object returned by one of the diagnostic plotting
functions when \code{patchwork} is not available.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a HmscEcoEvo plot list
}
