% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_history_tables.R
\name{standardize_history_tables}
\alias{standardize_history_tables}
\title{Standardize internal history tables}
\usage{
standardize_history_tables(proj, time_unit = "Ma")
}
\arguments{
\item{proj}{A \code{HmscEcoEvo_data} object.}

\item{time_unit}{Character; default "Ma".}
}
\value{
Updated \code{HmscEcoEvo_data} object.
}
\description{
Standardizes already stored history tables in A \code{HmscEcoEvo_data} object.
Missing optional probability columns are filled with 1. This function does
not infer historical processes.
}
