% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ecoevo.R
\name{validate_hmsc_ecoevo}
\alias{validate_hmsc_ecoevo}
\title{Validate a \code{hmsc_ecoevo} object}
\usage{
validate_hmsc_ecoevo(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Validate a \code{hmsc_ecoevo} object
}
